2-aiohttp/aiohttp_server/requirements.txt
aiohttp==3.7.4.post0
aiohttp-apispec==2.2.1
pyyaml==5.4.1
marshmallow~=3.13.0